Complete base JavaScript code + fix issues
Filled in operator event listener pseudo-code with actual code.
Fixed equals event listener and operator event listener so that num2 is
equal to what is displayed when the num1 and operator are known.
Fixed equals event listener so that operator equals undefined after an
operation is performed.
Fixed number event listener so that num1 can be multidigit, and num2 can
now also be multidigit because I addded a boolean variable that keeps
track of when num2 is beginning to be. written in the display.
Finally, comments were changed from pseudo-code
to minimal comments that help explain context