Fix a few bugs and add functionality to +/- button
Bug #1: An infinite number of numbers can be typed in the display. I was
able to add a limit of 9 numbers in the JavaScript.

Bug #2: Inner display grows larger as more numbers are typed in. To
solve this, I gave the inner display a max width in the css.

Finally, I added functionality to the sign change button. I added a new
element in the html that precedes the text element meant for the
display. The negative sign can now be turned on and off whenever, and
the operations take the signs into account.