Give functionality to teo more buttons + fix bugs
Gave functionality to the decimal button, percentage buttons, and added
keyboard support to the UI buttons. Fixed bug regarding solutions over-
flowing off of the display by converting numbers too large to scientific
notation, and by rounding numbers with decimals when the length of said
number is too large for the display. Finally, I fixed a bug where two
decimal places were able to be entered after a zero when zero was the
only number on display.