Fix bug
Found a bug where if the answer on display from a previous operation was
a negative, and the user subsequently would start typing in another
number, the negative would remain on screen. Accompanying this was a bug
in the logic of the code that would store the new num1 as the result
shown on the display, without taking into consideration the negative
shown on screen. This has been fixed